refactor(ui): share channel config extras and hint types
(cherry picked from commit dcf8308c8f969330b8864daf31ce5777c0dd97b8)
