+++ src/shared/config-ui-hints-types.ts
@@ -1,0 +1,13 @@
+export type ConfigUiHint = {
+  label?: string;
+  help?: string;
+  tags?: string[];
+  group?: string;
+  order?: number;
+  advanced?: boolean;
+  sensitive?: boolean;
+  placeholder?: string;
+  itemTemplate?: unknown;
+};
+
+export type ConfigUiHints = Record<string, ConfigUiHint>;
